#34758 Adds "More info" and "More fields" to the details panel.
The view of the currently-selected entity in the notes tab of the details panel now has a "More info" button. This will expand the widget and show additional fields that are otherwise hidden by default. In addition, the user may add fields to this view via a "More fields..." button that is active when "More info" has been checked on.

Closes #2.