#34582 and #34760 Version list context menu, tray selection, and events.
The version list view in the details panel now supports a right-click context menu that includes a standard set of actions that can be performed in RV on the current selection.

Selection of a version in the tray will now move the RV playhead if necessary.

RV events now handled for simulated GMA cut loading now in place, plus fixes for sorting and some UI polish.

Closes #3.