rv_activity_mode.py/resources: Added cuts btn to RV bottom toolbar
Added a cuts clapper button to the left hand side of the RV bottom
toolbar. Updated the resources to include the new clapper icon. There is
a sample listener method called "sample_cuts_action_listener" for
whomever wires up the button, as well as a pair of show/enable methods
for the button.